chore: move jwt models into auth/jwt
